add conversion to Polynomial from Poly
