+++ test/deprecated.jl
@@ -49,3 +49,16 @@ end
     PQexp = Pade(a, 8, 8)
     @test_deprecated padeval(PQexp, 1.0)
 end
+
+@testset "Converts to Polynomial" begin
+    p1 = Poly([1, 2, 3])
+    p = convert(Polynomial, p1)
+    @test p.coeffs == p1.coeffs
+    @test p.var == p1.var
+
+    p = convert(Polynomial{Float64}, p1)
+    @test p.coeffs == Float64[1, 2, 3]
+    @test p.var == p1.var
+
+    @test p1 + p == Polynomial([2, 4, 6])
+end
